Fixes #4 It was Skype applescript API's problem, in which I have to strip off those parenthesis myself before handling to Skype API interface.
This update moves the workflow to version 3.0. A big upgrade. Now you can simply type this in your Alfred:

call TARGET

where TARGET could be:
1. Skype username
2. phone number
3. Contact details stored in your Apple Contacts: first/middle/last names or corresponding phonetic names, nicknames, organisations

When you are typing the TARGET, a list of fuzzy matching results from Apple Contacts will also be shown to help you quickly select the right one. Even more, if the Skype is already open, the result entry corresponding to an online Skype username will be shown as green color to indicate the user is online. Hope you like it. Enjoy Skyping!